requirements.txt
instagrapi~=2.1.2
pillow~=10.3.0
requests~=2.32.2